+++ database/migrations/2025_01_10_130909_create_order_items_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('order_id');
             $table->foreignId('product_id');
-            $table->unsignedInteger('price');
-            $table->unsignedInteger('quantity');
+            $table->decimal('price', 10, 2);
+            $table->decimal('quantity', 10, 2);
             $table->timestamps();
         });
     }
